core: controlled-stop print uses %lld for NumCurrentTiStep (codex fixup)
Codex review of e469b827 flagged that the controlled-stop printout
cast All.NumCurrentTiStep to int and used %d, while the rest of the
codebase prints it as long long. Not a CG-stuck issue, just avoids
silent truncation on long cosmological runs whose sync-point counter
exceeds INT_MAX. Same Mac np=2 hernquist_sidm smoke as e469b827:
"sync-point 1275, Time=0.0219401" matches the prior receipt exactly.

+++ core/run.cc
@@ -156,12 +156,12 @@ void run(void)
             if(ThisTask == 0) {
                 const char *r = gizmo_controlled_stop_local_reason();
                 printf("Controlled stop (code=%d): %s. Leaving run loop at "
-                       "sync-point %d, Time=%g. No restart written at the "
+                       "sync-point %lld, Time=%g. No restart written at the "
                        "error point -- recover from an earlier periodic "
                        "restart.\n",
                        gizmo_controlled_stop_code(),
                        r ? r : "(reason set on other rank only)",
-                       (int)All.NumCurrentTiStep, All.Time);
+                       (long long)All.NumCurrentTiStep, All.Time);
                 fflush(stdout);
             }
             break;
